MantaPay V1 for Dolphin V3 deployment (#890)
* wip: start upgrading palllet


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: commit save point


* wip: save point


* Fix mantapay mock


* wip: checkpoint


* wip: checkpoint


* wip: checkpoint


* chore: fix signature error and add address partition


* wip: fix merge conflicts


* v1 to v2


* v1 to v2 everywhere


* disable benchmarkiing for asset-manager


* Use unsafe_no_prove_full in coin generation


* More features for precompute_coins


* proof-length update and comment


* Update for v2 circuit


* Clean up


* Fix manta-pay benchmarks


* Re-enable asset-manager benchmarks


* Re-enable asset-manager benchmarks


* Clean up


* Use latest comimt of manta-rs


* Add dolphin-v3 chainspec


* Fix clippy


* Fix clippy for manta-pay tests


* Taplo fmt


* More clippy


* More consistent chain-spec string


* Simplify 0 check in id conversions


* Make utxo.is_transparent an enum isntead of bool, with Opaque variant as default


* Fix clippy


* bug fixed:


* Update manta-rs to insert receiver utxo in to-public precomputed coins. Compiles but InvalidAuthorizationSignature


* Use dolphin_properties for the new chain-spec, not calamari


* rename is_transparent field to transparency since no longer a bool


* Update with LightIncomingNote


* Revert precompute_coins.rs


* Remove println! macros


* Update accumulator in to_private::prove_full and add a new test


* Add 2 more nodes for calamari-local integration tests


* Clean up tests


* Remove erroneously pulled in Cargo.lock changes


* Fix fmt in benchmark


* Fix precompute_coins.rs


* Update to circuits v3


* Fix decode error


* insert receiver1 before receiver 0 utxo


* Fix benchmark tests


* Fix linters


* Clean up types and add comments


* BUmp rust version to 1.62


* Error handle a few expects


* Remove unwraps


* Update manta-rs commit


* Update Dolphin weights


* Fix linter


* Use latest manta-rs commit|


* 3 nodes instead of 5 in chain-spec


* Use json genesis file for DolphinV3


* Remove unused imports


* Update genesis json

+++ primitives/manta/src/types.rs
@@ -45,7 +45,7 @@ pub type AccountIndex = u32;
 pub type CalamariAssetId = u128;
 
 /// Dolphin Asset Id Type
-pub type DolphinAssetId = u32;
+pub type DolphinAssetId = u128;
 
 /// Balance of an Account
 pub type Balance = u128;
